backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
distro==1.9.0
fastapi==0.128.0
greenlet==3.3.0
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
joblib==1.5.3
lxml==6.0.2
numpy==2.4.1
packaging==25.0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
pluggy==1.6.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pypdfium2==5.3.0
pytest==9.0.2
python-docx==1.2.0
python-multipart==0.0.21
reportlab==4.4.9
scikit-learn==1.8.0
scipy==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.45
starlette==0.50.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
